backend/requirements.txt
fastapi
uvicorn[standard]
numpy
torch
torchaudio
torchvision
pillow
transformers
silero-vad
SpeechRecognition
pydub
omegaconf
soundfile
sqlite3
requests
langchain_google_genai
langchain
google-generativeai
langchain-google-genai
pillow
transformers
dotenv
hf_xet
torchvision